+++ moPepGen/svgraph/VariantPeptideDict.py
@@ -466,7 +466,7 @@ class MiscleavedNodes():
                 cur_metadata.has_variants = bool(cur_variants)
 
                 if is_valid or is_valid_start:
-                    cur_nodes = []
+                    cur_nodes:List[PVGNode] = []
                     cut_offset = sec.location.start
                     for node in nodes:
                         if cut_offset == 0:
@@ -474,7 +474,7 @@ class MiscleavedNodes():
                             continue
                         if len(node.seq.seq) > cut_offset:
                             node = node.copy()
-                            node.truncate_left(cut_offset)
+                            node.truncate_right(cut_offset)
                             cur_nodes.append(node)
                             cut_offset = 0
                         else: